+++ src/ContaoManager/Plugin.php
@@ -10,14 +10,25 @@ use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Config\ConfigInterface;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
 use Hofff\Contao\Consent\Bridge\HofffContaoConsentBridgeBundle;
+use MadeYourDay\RockSolidCustomElements\RockSolidCustomElementsBundle;
 
 final class Plugin implements BundlePluginInterface
 {
-    /** @return ConfigInterface[] */
+    /**
+     * @return ConfigInterface[]
+     *
+     * @psalm-suppress UndefinedClass
+     */
     public function getBundles(ParserInterface $parser) : array
     {
-        return [BundleConfig::create(HofffContaoConsentBridgeBundle::class)
-                ->setLoadAfter([ContaoCoreBundle::class]),
+        return [
+            BundleConfig::create(HofffContaoConsentBridgeBundle::class)
+                ->setLoadAfter(
+                    [
+                        ContaoCoreBundle::class,
+                        RockSolidCustomElementsBundle::class,
+                    ]
+                ),
         ];
     }
 }