+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 
 # Changelog
 
+## [Unreleased]
+
+## [1.2.0] - (2020-08-07)
+
+### Added
+
+- Add basic support for rocksolid custom elements
+- Add configuration option to adjust supported content elements and frontend modules
+
+### Fixed
+
+- Fix license information in the composer.json
+
 ## [1.1.1] - (2020-05-14)
 
 ### Fixed
@@ -14,5 +27,6 @@
 - Add info messages if no consent tag is assign for supported frontend modules and content elements
 
 
+[1.2.0]: https://github.com/hofff/contao-consent-bridge/compare/1.0.0...1.1.0
 [1.1.1]: https://github.com/hofff/contao-consent-bridge/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/hofff/contao-consent-bridge/compare/1.0.0...1.1.0